feat: Add fetchAllModulosCompleteData function to MainProvider
The fetchAllModulosCompleteData function is added to the MainProvider component in order to fetch the complete data for all modules. This function makes a GET request to the "/api/modules" endpoint with the user's token as the authorization header. If the request is successful, the data is returned. If there is an error, a toast notification is displayed. This function is used to load the list of modules.