feat: Update ListaCursos component to display formatted dates and times
The ListaCursos component in the Admin/SubModules directory is updated to display formatted dates and times for the curso object. This includes converting the fecha_inicio and fecha_fin properties to a localized date format and the horarioInicio and horarioFin properties to a localized time format. This improves the readability and user experience of the component.