refactor: Update AdminModuleNavbar component to include deleted modules tab
The AdminModuleNavbar component in the client/src/components/Admin/AdminModuleNavbar.jsx file is updated to include a new tab for displaying deleted modules. This allows users to easily access and manage deleted modules within the admin interface. The selectedTab state is also updated to include the "deletedModules" option. This enhancement improves the functionality and user experience of the AdminModuleNavbar component.