feat: Add LoaderAE component to PageRegister
Add the LoaderAE component to the PageRegister component in the client/src/pages/PageRegister.jsx file. This component will be used to display a loading indicator while the steps of the registration process are being loaded. This feature enhances the user experience by providing visual feedback during the loading process.