refactor: Add @radix-ui/react-checkbox dependency and Checkbox component
Add the "@radix-ui/react-checkbox" dependency to the package.json and package-lock.json files. This change ensures that the project has the necessary dependencies to use the Checkbox component. The Checkbox component is also added to the project, providing a customizable checkbox UI element for use in forms and user interfaces.